requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
bleach==6.2.0
blinker==1.9.0
boto3==1.35.55
botocore==1.35.55
Brotli==1.1.0
cachetools==5.5.0
category-encoders==2.6.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
cmake==3.30.5
cohere==5.11.4
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cssselect2==0.7.0
cvzone==1.6.1
cycler==0.12.1
dask==2024.11.2
dataclasses-json==0.6.7
datasets==3.1.0
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.5
fastavro==1.9.7
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
gdown==5.2.0
geopandas==1.0.1
gitdb==4.0.11
github3.py==4.0.1
GitPython==3.1.43
google-auth==2.36.0
greenlet==3.1.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
imutils==0.5.4
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
jellyfish==1.1.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
kaleido==0.2.1
kiwisolver==1.4.7
lida==0.0.14
llama-index-agent-openai==0.4.0
llama-index-core==0.12.0
llama-index-llms-llama-api==0.3.0
llama-index-llms-openai==0.3.0
llama-index-program-openai==0.3.0
llamaapi==0.1.36
llmx==0.0.21a0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
matplotlib-venn==1.1.1
mdurl==0.1.2
mizani==0.13.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
narwhals==1.14.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
openai==1.54.5
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pillow==10.2.0
platformdirs==4.3.6
plotly==5.24.1
plotnine==0.14.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydyf==0.11.0
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyogrio==0.10.0
pyparsing==3.2.0
pyphen==0.15.0
pyproj==3.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
python-slugify==8.0.4
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==65.5.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
streamlit==1.40.1
streamlit-card==1.0.2
sympy==1.13.1
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.20.3
toml==0.10.2
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typer==0.13.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
watchdog==6.0.0
wcwidth==0.2.13
weasyprint==63.0
xarray==2024.9.0
xlrd==2.0.1
yellowbrick==1.5
zipp==3.20.2
